tui: move the package screens into tui/packages.py
Seven screens share one set of helpers, so no single closure is the module
and the union of the seven is: six names cross outward and none inward.
