+++ tests/unit/test_i18n.py
@@ -107,10 +107,9 @@ def in_a_table() -> set[str]:
     from gentoo_install.model.mirrors import GENTOO_SITES, GENTOOZH_SITES
     from gentoo_install.plan.system import LOGGERS
     from gentoo_install.tui.mirror import BINHOSTS, GENTOOZH_CHANNELS, SYNC_METHODS
+    from gentoo_install.tui.packages import DISPLAY_MANAGERS, GRAPHICS
     from gentoo_install.tui.screens import (
         RAM_SHARES,
-        DISPLAY_MANAGERS,
-        GRAPHICS,
         INSTALL_MODES,
         KERNELS,
         LICENSES,
